refactor(analysis): group decision_*.py into an analysis/decisions/ package
Collect all five decision modules into one dedicated folder:
  analysis/decisions/
    extractor.py        (former decision_extractor.py)
    prompts.py          (LLM prompt strings carved out of the extractor)
    evolution.py        (former decision_evolution.py)
    gate.py             (former decision_gate.py)
    provenance.py       (former decision_provenance.py)
    semantic_match.py   (former decision_semantic_match.py)

Intra-package imports now point at repowise.core.analysis.decisions.*, and
the five old analysis/decision_*.py paths are thin re-export shims, so every
existing importer (cli, crud, orchestrator/persist, page_generator, server
tool_why, tests) keeps working unchanged.

The 1700-line extractor only gets the safe prompts carve-out here; the
per-source method split + extract_all pluggable-extractor refactor + the
3x LLM-batch dedup are intentionally left for a follow-up (DecisionExtractor
is one tightly-coupled class).

No behavior change. Ruff clean, no new mypy errors; tests/unit green
(2383 passed).
