+++ packages/core/src/repowise/core/analysis/decisions/__init__.py
@@ -1,0 +1,4 @@
+"""Architectural decision intelligence — extraction, provenance, gating,
+semantic dedup, and evolution. Grouped from the former analysis/decision_*.py
+modules.
+"""
